refactor: change pfr/utils.py to pfr/pfr_utils.py and update respective test names to ensure module uniqueness
