+++ tests/scraping/pfr/test_pfr_utils.py
@@ -1,17 +1,17 @@
 import pytest
 from bs4 import BeautifulSoup
-from scrape_and_score.scraping.pfr import utils
+from scrape_and_score.scraping.pfr import pfr_utils
 from unittest.mock import patch
 
 
 def test_get_last_name_first_initial():
-    assert utils.get_last_name_first_initial("Tom Brady") == "B"
+    assert pfr_utils.get_last_name_first_initial("Tom Brady") == "B"
     with pytest.raises(Exception):
-        utils.get_last_name_first_initial("Brady")
+        pfr_utils.get_last_name_first_initial("Brady")
 
 
 def test_check_name_similarity():
-    similarity = utils.check_name_similarity("Tom Brady", "Tom Brady")
+    similarity = pfr_utils.check_name_similarity("Tom Brady", "Tom Brady")
     assert similarity == 100
 
 
@@ -21,7 +21,7 @@ def test_order_players_by_last_name():
         {"player_name": "Aaron Rodgers"},
         {"player_name": "Josh Allen"},
     ]
-    result = utils.order_players_by_last_name(players)
+    result = pfr_utils.order_players_by_last_name(players)
     assert len(result["B"]) == 1
     assert len(result["R"]) == 1
     assert len(result["A"]) == 1
@@ -30,7 +30,7 @@ def test_order_players_by_last_name():
 def test_calculate_distance():
     city1 = {"latitude": 40.7128, "longitude": -74.0060}  # NYC
     city2 = {"latitude": 34.0522, "longitude": -118.2437}  # LA
-    dist = utils.calculate_distance(city1, city2)
+    dist = pfr_utils.calculate_distance(city1, city2)
     assert 2400 < dist < 3000
 
 
@@ -38,7 +38,7 @@ def test_parse_team_totals():
     html = '<table><tfoot><tr><td data-stat="xpa">10</td><td data-stat="xpm">8</td></tr></tfoot></table>'
     soup = BeautifulSoup(html, "html.parser")
     tfoot = soup.find("tfoot")
-    result = utils.parse_team_totals(tfoot)
+    result = pfr_utils.parse_team_totals(tfoot)
     assert result == {"team_total_xpa": "10", "team_total_xpm": "8"}
 
 
@@ -48,7 +48,7 @@ def test_filter_metrics_by_week():
         {"week": "1", "yards": 200},
         {"week": "2", "yards": 150},
     ]
-    filtered = utils.filter_metrics_by_week(metrics)
+    filtered = pfr_utils.filter_metrics_by_week(metrics)
     assert len(filtered) == 2
     assert filtered[0]["week"] == "1"
     assert filtered[1]["week"] == "2"
@@ -68,16 +68,16 @@ def test_is_team_game_log_modified():
                "fourth_down_att": 2, "penalties": 4, "penalty_yds": 35,
                "fmbl_lost": 1, "interceptions": 1, "turnovers": 2, "time_of_poss": 29.5}
     persisted = current.copy()
-    assert utils.is_team_game_log_modified(current, persisted) is False
+    assert pfr_utils.is_team_game_log_modified(current, persisted) is False
     current["points_for"] = 28
-    assert utils.is_team_game_log_modified(current, persisted) is True
+    assert pfr_utils.is_team_game_log_modified(current, persisted) is True
 
 
 def test_get_additional_metrics_qb():
-    fields = utils.get_additional_metrics("QB")
+    fields = pfr_utils.get_additional_metrics("QB")
     assert "cmp" in fields and isinstance(fields["cmp"], list)
 
 
 def test_get_additional_metrics_invalid():
     with pytest.raises(Exception):
-        utils.get_additional_metrics("K")
+        pfr_utils.get_additional_metrics("K")
